Move Environment and Test config to config.py.
